Added support for retrieving data on retired players.
Added a flag to tell the script to grab retired players.  Also set some
defaults for the hashes when collecting individual rows of data.

Also added a test case for a really old player whose season is not
linkable.  This tripped the script up because we’re expecting each
season to be linked.