Fixes for salary importing
Made the file variable a class attribute so it could be set in a unit
test (instead of having to be a CL arg).  Moved all database logic into
a try/finally so we could add more logging during errors.  Also, added
a null check on rg_position so we don’t pick up players that won’t
match.