Fix for issue 13.
Added a defense_vs_position table to cache all DvP calculations in
order to speed up regression.  Also added a model class
DefenseVsPosition and manager DefenseVsPositionManager to handle
database interactions.

Now, the Projections.calculate_defense_vs_position() call will
initially prime the cache to speed things up.  Also, the logic for
fetching the values has been moved to DefenseVsPositionManager, and if
there is no database hit, the value is calculated and then written to
the database.  This should be the last time the value needs to be
calculated for the stat/position/team/season/date combination.