Added depth chart creation
We now have two depth charts.  The first is a global team/position view
that incorporates what a team’s depth chart typically looks like - it
does not incorporate the injuries for the day.  The second is more of a
lookup map keyed off player id.  This provides the expected minutes for
a player, based on their average minutes as well as redistribution of
minutes from injured players, and their adjusted rank in the depth
chart.

This is part of issue 16.