Fix for issue 16 (Keep track of injuries)
The InjuryManager can now evaluate game logs for a season and detect
which games a player sat out (for any reason).  It’s also capable of
handling players that have been traded mid-season.

Additionally, there is a stand-alone script that invokes this and can
be run daily to keep up-to-date on past injuries.