Fix for issue 30 (Issues with overlapping injuries)
Added a function in InjuryManager that sweeps through all injuries and
determines, based on a player’s game logs, whether that injury is
correct.  If not, it gets corrected.
A change was also made to fix_injury_entry so all conditions are
checked instead of just having an else clause.
This change has been incorporated into the all_tasks script so it will
be run after all new injuries are established.