Fix for issue 30 (Issues with overlapping injuries)
Fixed issue with unread results in cursor.  This was happening because
we’d try to perform an update on an injury while there were still
cursor results.  The solution is to record all injury/game-date pairs
(game dates that fall during the alleged injury) and deal with them
once all results are read.

Also added duplicate removal.