Added scraping for yesterday's lineup.
Scraping yesterday’s lineup will pick up the stats for most of the
probable starters for today.  This lets us get going early in the
morning instead of having to do lots of scraping when line-ups come out
in the afternoon.

Also added a “sleep” CLI option to dictate how long HTTP calls should
sleep for.