Fixed logging of unfound player name.
Previously, a player whose name doesn’t map from Rotowire to BBR would
be logged, but with the overwritten player_name that is None.  I’ve
saved off the player_name parameter as original_name and am logging
with that, so if player_name gets overwritten with None, we can still
see who the player is.